refactor: made separate function for making malformed prizebond numbers into proper number
if the number is less than 7 in size and is generated, then it must be padded with 0s.
It was done before with the same code but inside the code, now it's factored out into a static method
So that it can be used in other parts of the codebase.